fix: password fields not synchronized between livewire components
Instead of taking Povillas approach, i ended up installing Breeze starter kit with livewire components instead of the default blade ones, and then i've faced the problem that there was a problem related to component communication: 'RegisterPasswords' and 'register' had both the password property - the commit doesn't show this, but all this started with getting an error of password property not defined, defining it but then facing a new and silent problem involving synchronization -, but they didn't shared any reference: password from 'RegisterPasswords' was making data binding happen (the password strength progress bar working etc.), but the password that was getting validated in 'register' wasn't any related to it. After exploring the possibilities, i concluded letting all the business logic related to the form in a single component and just encapsulate the password structure in a blade component rather than another livewire component would be the most suitable approach: the actions on the livewire component would be propagated to the blade component on re-rendering it. Considering the size of the component, nesting would bring more overhead than advantages: implementing a workaround to solve the defficiency of Livewire not out-of-the-box having how to "establish communication" between components through prop changes wouldn't be worthwhile. I've read the documentation: each component it's an island, so any change on the props of one component wouldn't affect at all any nesting one.